requirements.txt
asgiref==3.11.0
Django==6.0.1
django-debug-toolbar==6.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
pyparsing==3.3.2
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
